SettingsProvider useSettings refactoring
