+++ src/hooks/useSettings.ts
@@ -1,0 +1,18 @@
+import { useContext } from 'react';
+
+import { Settings } from 'api/useGetSettings';
+import { SettingsContext } from 'providers/SettingsProvider';
+
+/**
+ * The `useSettings` hook returns the current `SettingsContext` value.
+ * @returns {Settings} The current `SettingsContext` value, `Settings`.
+ * @see {@link Settings}
+ */
+export const useSettings = (): Settings => {
+  const context = useContext(SettingsContext);
+  if (!context) {
+    throw new Error('useSettings hook must be used within a SettingsContextProvider');
+  }
+
+  return context;
+};
